+++ packages/_internals/voice/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @internal/voice
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies:
+  - @internal/ai-sdk-v5@0.0.63
+
 ## 0.0.15
 
 ### Patch Changes